requirements_streamlit.txt
streamlit>=1.28.0
pandas>=2.0.0
qrcode>=7.4.2
Pillow>=10.0.0
requests>=2.31.0
python-dateutil>=2.8.2
plotly>=5.14.0